// src/components/Flex/types.ts
export interface IFlexProps {
    /**
     * direction of display flex. Accept breakpoints: md;lg
     */
    direction?: string;

    /**
     * Accept breakpoints: md;lg
     */
    wrap?: string;

    /**
     * Accept breakpoints: md;lg
     */
    flow?: string;
    
    /**
     * Accept breakpoints: md;lg
     */
    align?: string;
    
    /**
     * Accept breakpoints: md;lg
     */
    justify?: string;
    
    /**
     * Accept breakpoints: md;lg
     */
    grow?: string;
    
    /**
     * Accept breakpoints: md;lg
     */
    shrink?: string;
    
    /**
     * Accept breakpoints: md;lg
     */
    gap?: string;
    
    /**
     * Accept breakpoints: md;lg
     */
    w?: string;
    
    /**
     * Accept breakpoints: md;lg
     */
    h?: string;

    padding?: string;
    
    margin?: string;
}